Refactors StatsWriter into its own package
